set file viewer from command line

+++ scripts/osx/install-deps.sh
@@ -4,29 +4,32 @@
 STARTING_DIR=$(pwd)
 TARGET_DIR=~/Projects/yhat/rodeo
 NODE_VERSION=6
+CACHE_MIN=999999999
 
 #guarantee nvm
 echo '#guarantee nvm'
-brew install nvm
+if [ -s "$NVM_DIR/nvm.sh" ]; then
+    brew install nvm
+fi
 source $(brew --prefix nvm)/nvm.sh
 
 #install script dependencies
 echo '#install script dependencies'
 nvm install $NODE_VERSION
-npm install -g gulp-cli electron-prebuilt
+npm install --cache-min $CACHE_MIN -g gulp-cli electron-prebuilt
 
 #build application (production and dev dependnencies)
 cd $TARGET_DIR
 echo '#build application (production and dev dependnencies)'
 rm -rf app dist
-npm install
+npm install --cache-min $CACHE_MIN
 gulp build
 cd $STARTING_DIR
 
 #install application including production dependencies only (no dev)
 echo '#install application including production dependencies only (no dev)'
 cd "$TARGET_DIR/app"
-npm install --production
+npm install --production --cache-min $CACHE_MIN
 cd $STARTING_DIR
 
 #check brew
